Stacked architecture setup
