+++ lib/app/app.dart
@@ -1,0 +1,18 @@
+import 'package:hng/ui/home/home_view.dart';
+import 'package:hng/ui/login/login_view.dart';
+import 'package:stacked/stacked_annotations.dart';
+import 'package:stacked_services/stacked_services.dart';
+
+@StackedApp(
+  routes: [
+    MaterialRoute(page: LoginView),
+    MaterialRoute(page: HomeView),
+  ],
+  dependencies: [
+    LazySingleton(classType: NavigationService),
+  ],
+)
+class AppSetup {
+  /* This class serves no other purpose other than having the
+   StackedApp annotation attached to it*/
+}
