+++ lib/app/app.router.dart
@@ -1,0 +1,47 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// StackedRouterGenerator
+// **************************************************************************
+
+// ignore_for_file: public_member_api_docs
+
+import 'package:flutter/material.dart';
+import 'package:stacked/stacked.dart';
+
+import '../ui/home/home_view.dart';
+import '../ui/login/login_view.dart';
+
+class Routes {
+  static const String loginView = '/login-view';
+  static const String homeView = '/home-view';
+  static const all = <String>{
+    loginView,
+    homeView,
+  };
+}
+
+class StackedRouter extends RouterBase {
+  @override
+  List<RouteDef> get routes => _routes;
+  final _routes = <RouteDef>[
+    RouteDef(Routes.loginView, page: LoginView),
+    RouteDef(Routes.homeView, page: HomeView),
+  ];
+  @override
+  Map<Type, StackedRouteFactory> get pagesMap => _pagesMap;
+  final _pagesMap = <Type, StackedRouteFactory>{
+    LoginView: (data) {
+      return MaterialPageRoute<dynamic>(
+        builder: (context) => LoginView(),
+        settings: data,
+      );
+    },
+    HomeView: (data) {
+      return MaterialPageRoute<dynamic>(
+        builder: (context) => HomeView(),
+        settings: data,
+      );
+    },
+  };
+}
