--- lib/app_routing/route_names.dart
@@ -1,19 +1,0 @@
-//ROUTE NAMES
-
-//* Launch and Onboarding
-const splashViewRoute = '/splash';
-const welcomeViewRoute = '/welcome';
-
-//* Authentication
-const logInViewRoute = '/signIn';
-const registerViewRoute = '/register';
-const emailRegisterViewRoute = '/register/email';
-
-//* Home
-const homeViewRoute = '/home';
-
-//* Notifications
-
-//* Search
-
-//* Settings
