+++ lib/main.dart
@@ -1,20 +1,24 @@
 import 'package:flutter/material.dart';
+import 'package:hng/app/app.locator.dart';
+import 'package:stacked_services/stacked_services.dart';
 
-import 'app_routing/app_navigator.dart';
-import 'app_routing/app_router.dart';
-import 'views/home/home_view.dart';
+import 'app/app.router.dart';
 
-void main() => runApp(MyApp());
+void main() {
+  setupLocator();
+  runApp(MyApp());
+}
 
 // ignore: use_key_in_widget_constructors
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      navigatorKey: AppNavigator.key,
-      onGenerateRoute: AppRouter.generateRoute,
+      debugShowCheckedModeBanner: false,
+      navigatorKey: StackedService.navigatorKey,
+      onGenerateRoute: StackedRouter().onGenerateRoute,
       title: 'ZuriChat App',
-      home: const HomeView(),
+      initialRoute: Routes.homeView,
     );
   }
 }
