+++ lib/ui/home/home_viewmodel.dart
@@ -1,0 +1,3 @@
+import 'package:stacked/stacked.dart';
+
+class HomeViewModel extends BaseViewModel {}
