+++ lib/ui/home/widgets/custom_plugin_list_tile.dart
