+++ lib/ui/login/login_viewmodel.dart
@@ -1,0 +1,3 @@
+import 'package:stacked/stacked.dart';
+
+class LoginViewModel extends BaseViewModel {}
