--- lib/view_models/dm_view_model.dart
@@ -1,1 +1,0 @@
-
