+++ pubspec.yaml
@@ -16,9 +16,15 @@ dependencies:
   intl: ^0.17.0
   url_launcher: ^6.0.9
 
+  # Stacked
+  stacked: ^2.1.8
+  stacked_services: ^0.8.6
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  stacked_generator: ^0.4.6
+  build_runner:
 
 flutter:
   uses-material-design: true
